[Rename] 컴포넌트명, 폴더 위치 정리
