+++ src/components/wrapper/PostForm.js
@@ -2,10 +2,10 @@ import { useState, useCallback, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { useNavigate } from 'react-router';
 
+import TagBox from '@components/wrapper/TagBox';
+import Header from '@pages/main/Header';
 import StyledIndex from '@styles/main/Index-styled';
 import StyledCreate from '@styles/main/post/Create-styled';
-import Header from '@pages/main/Header';
-import TagBox from '@pages/main/post/create-post/TagBox';
 
 const PostForm = ({ post, onSubmit }) => {
   const [enteredCategory, setEnteredCategory] = useState('');
